requirements.txt
Django
psycopg2
djangorestframework
pandas
sqlalchemy
chardet